requirements.txt
fastapi
uvicorn
sqlmodel
pymysql
httpx
python-dotenv
cryptography
PyPDF2
python-docx
openpyxl
python-multipart
pandas
matplotlib
python-jose[cryptography]
passlib[bcrypt]